moved _message.py
