+++ pyodsp/dec/node/_message.py
